+++ src/modules/page/PageAccessToAudience.tsx
@@ -11,18 +11,24 @@ import { CHANGE_PAGE_ACCESS_PRIVILEGES_LIST } from 'modules/page/pages.constants
 import { PageAccessPrivilegesType, PageAccessToAudiencesPropsType } from 'modules/page/pages.types';
 import RemoveFromTeamPopup from 'modules/people/RemoveFromTeamPopup';
 import Image from 'next/image';
+import { accessPermissionForPage } from 'utils/accessPermission/accessPermission';
+import { PERMISSION_MESSAGES } from 'utils/accessPermission/accessPermission.constants';
+import { PERMISSION_TYPES } from 'utils/accessPermission/accessPermission.types';
+import { checkIfCurrentUser } from 'utils/accessPermission/accessPermission.utils';
 import { convertEmailUsernameToName, getUserNameFromEmail } from 'utils/common';
 import AsyncDropdown from 'components/asyncDropdown/AsyncDropdown';
 import Avatar from 'components/common/avatar';
 import { toast } from 'components/common/toast/Toast';
+import { TOAST_MESSAGES } from 'components/common/toast/toast.constants';
 
 const PageAccessToAudiences: FC<PageAccessToAudiencesPropsType> = ({
   resource_type,
   privilege,
   pageId,
   resource_audience_id,
-  resource_audience_type,
   user,
+  userPrivilege,
+  resource_audience_type,
 }) => {
   const dropdownRef = useRef<HTMLDivElement>(null);
   const role = CHANGE_PAGE_ACCESS_PRIVILEGES_LIST.find((role) => role.value === privilege);
@@ -33,7 +39,9 @@ const PageAccessToAudiences: FC<PageAccessToAudiencesPropsType> = ({
   const { refetch: refetchAudiencesByPageId } = useGetAudiencesByPageIdQuery({ pageId }, { skip: !pageId });
   const [changeRole] = usePatchChangeAudienceRoleInPageMutation();
   const [deleteAudience] = useDeleteAudienceFromPageAccessMutation();
+  const checkIfUser = checkIfCurrentUser(user?.email ?? '');
   const userName = convertEmailUsernameToName(getUserNameFromEmail(user?.email || resource_audience_type));
+  const checkPermission = accessPermissionForPage(userPrivilege);
 
   const handleOpenChangeRoleDropdown = () => {
     setOpenChangeRoleDropdown(true);
@@ -43,25 +51,31 @@ const PageAccessToAudiences: FC<PageAccessToAudiencesPropsType> = ({
     setOpenChangeRoleDropdown(false);
   };
 
-  const handleRoleChange = (selectedOption: PageAccessPrivilegesType) => {
-    changeRole({
-      pageId: pageId,
-      body: {
-        audience_id: resource_audience_id,
-        role: selectedOption?.value,
-      },
-    })
-      .unwrap()
-      .then(() => {
-        setSelectedRole(selectedOption);
-        setOpenChangeRoleDropdown(false);
-        setIsHoveredDropdown(false);
-        refetchAudiencesByPageId();
-        toast.success('Role changed successfully');
+  const handleRoleChange = async (selectedOption: PageAccessPrivilegesType) => {
+    if (!checkPermission) {
+      toast.error(PERMISSION_MESSAGES[PERMISSION_TYPES.ROLE_CHANGE]);
+
+      return;
+    } else {
+      await changeRole({
+        pageId: pageId,
+        body: {
+          audience_id: resource_audience_id,
+          role: selectedOption?.value,
+        },
       })
-      .catch((err) => {
-        toast.error(err?.data?.error || 'Failed to change role');
-      });
+        .unwrap()
+        .then(() => {
+          setSelectedRole(selectedOption);
+          setOpenChangeRoleDropdown(false);
+          setIsHoveredDropdown(false);
+          refetchAudiencesByPageId();
+          toast.success(TOAST_MESSAGES.SUCCESS_AUDIENCE_ROLE_CHANGED);
+        })
+        .catch((err) => {
+          toast.error(err?.data?.error || TOAST_MESSAGES.FAILED_AUDIENCE_ROLE_CHANGED);
+        });
+    }
   };
 
   const handleOpenRemoveFromTeamPopup = () => {
@@ -72,30 +86,36 @@ const PageAccessToAudiences: FC<PageAccessToAudiencesPropsType> = ({
     setIsOpenRemoveFromTeamPopup(false);
   };
 
-  const handleDeleteAudience = () => {
-    deleteAudience({
-      pageId: pageId,
-      body: {
-        audience_id: resource_audience_id,
-      },
-    })
-      .unwrap()
-      .then(() => {
-        handleCloseRemoveFromTeamPopup();
-        refetchAudiencesByPageId();
-        toast.success('Audience deleted successfully');
+  const handleDeleteAudience = async () => {
+    if (!checkPermission) {
+      toast.error(PERMISSION_MESSAGES[PERMISSION_TYPES.ROLE_CHANGE]);
+
+      return;
+    } else {
+      await deleteAudience({
+        pageId: pageId,
+        body: {
+          audience_id: resource_audience_id,
+        },
       })
-      .catch((err) => {
-        handleCloseRemoveFromTeamPopup();
-        toast.error(err?.data?.error || 'Failed to delete audience');
-      });
+        .unwrap()
+        .then(() => {
+          handleCloseRemoveFromTeamPopup();
+          refetchAudiencesByPageId();
+          toast.success(TOAST_MESSAGES.SUCCESS_AUDIENCE_DELETED);
+        })
+        .catch((err) => {
+          handleCloseRemoveFromTeamPopup();
+          toast.error(err?.data?.error || TOAST_MESSAGES.FAILED_AUDIENCE_DELETED);
+        });
+    }
   };
 
   useOnClickOutside(dropdownRef, handleCloseChangeRoleDropdown);
 
   return (
     <>
-      <div className='f-12-400 px-2 bg-white flex justify-between items-center'>
+      <div className='f-12-400 pl-2 bg-white flex justify-between items-center'>
         <div className='flex items-center justify-start'>
           <div className='flex items-start justify-start gap-x-1 w-[140px]'>
             <>
@@ -106,7 +126,10 @@ const PageAccessToAudiences: FC<PageAccessToAudiencesPropsType> = ({
                   className='w-4 h-4 rounded-full text-white f-8-400 flex items-center justify-center'
                 />
               </div>
-              <span>{userName}</span>
+              <div className='flex justify-center items-center gap-1'>
+                {userName}
+                <span className='f-12-400 text-GRAY_700'>{checkIfUser && '(You)'}</span>
+              </div>
             </>
           </div>
           <span className='flex text-wrap flex-wrap break-words whitespace-normal items-center justify-start gap-1 w-[100px]'>
@@ -115,20 +138,25 @@ const PageAccessToAudiences: FC<PageAccessToAudiencesPropsType> = ({
           </span>
         </div>
 
-        <AsyncDropdown
-          onOpen={handleOpenChangeRoleDropdown}
-          onClose={handleCloseChangeRoleDropdown}
-          isOpen={openChangeRoleDropdown}
-          onDelete={handleOpenRemoveFromTeamPopup}
-          onChange={(role) => handleRoleChange(role as PageAccessPrivilegesType)}
-          options={CHANGE_PAGE_ACCESS_PRIVILEGES_LIST}
-          selectedValue={selectedRole}
-          defaultValue={role as PageAccessPrivilegesType}
-          showDelete
-          isHoveredDropdown={isHoveredDropdown}
-          setIsHoveredDropdown={setIsHoveredDropdown}
-          isOverflowStyle
-        />
+        {checkPermission ? (
+          <AsyncDropdown
+            onOpen={handleOpenChangeRoleDropdown}
+            onClose={handleCloseChangeRoleDropdown}
+            isOpen={openChangeRoleDropdown}
+            onDelete={handleOpenRemoveFromTeamPopup}
+            onChange={(role) => handleRoleChange(role as PageAccessPrivilegesType)}
+            options={CHANGE_PAGE_ACCESS_PRIVILEGES_LIST}
+            selectedValue={selectedRole}
+            defaultValue={role as PageAccessPrivilegesType}
+            showDelete
+            showSelectedIcon
+            isHoveredDropdown={isHoveredDropdown}
+            setIsHoveredDropdown={setIsHoveredDropdown}
+            isOverflowStyle
+          />
+        ) : (
+          <span className='flex justify-between items-start f-12-400 text-GRAY_1000 pl-4 py-3 pr-2'>{role?.label}</span>
+        )}
       </div>
       <RemoveFromTeamPopup
         isOpen={isOpenRemoveFromTeamPopup}
