+++ src/modules/people/components/teamMembers/TeamMembersRole.tsx
@@ -10,8 +10,12 @@ import { TEAM_MEMBERS_PRIVILEGES_LIST } from 'modules/people/people.constants';
 import { TeamMemberAccessPrivilegesType, TeamMembersRolePropsType } from 'modules/people/people.types';
 import RemoveFromTeamPopup from 'modules/people/RemoveFromTeamPopup';
 import { RootState } from 'store';
+import { accessPermissionForPeople } from 'utils/accessPermission/accessPermission';
+import { PERMISSION_MESSAGES } from 'utils/accessPermission/accessPermission.constants';
+import { PERMISSION_TYPES } from 'utils/accessPermission/accessPermission.types';
 import AsyncDropdown from 'components/asyncDropdown/AsyncDropdown';
 import { toast } from 'components/common/toast/Toast';
+import { TOAST_MESSAGES } from 'components/common/toast/toast.constants';
 
 const TeamMembersRole: FC<TeamMembersRolePropsType> = ({ value }) => {
   const { user_id, privilege } = value;
@@ -30,6 +34,7 @@ const TeamMembersRole: FC<TeamMembersRolePropsType> = ({ value }) => {
     { skip: !organizationId },
   );
   const dropdownRef = useRef<HTMLDivElement>(null);
+  const checkPermission = accessPermissionForPeople();
 
   const handleOpenChangeRoleDropdown = () => {
     setOpenChangeRoleDropdown(true);
@@ -40,24 +45,30 @@ const TeamMembersRole: FC<TeamMembersRolePropsType> = ({ value }) => {
   };
 
   const handleRoleChange = (selectedOption: TeamMemberAccessPrivilegesType) => {
-    changeRole({
-      organizationId: organizationId,
-      body: {
-        user_id: user_id,
-        role: selectedOption?.value,
-      },
-    })
-      .unwrap()
-      .then(() => {
-        setSelectedRole(selectedOption);
-        setOpenChangeRoleDropdown(false);
-        setIsHoveredDropdown(false);
-        refetchAudiencesByOrganizationId();
-        toast.success('Role changed successfully');
+    if (!checkPermission) {
+      toast.error(PERMISSION_MESSAGES[PERMISSION_TYPES.ROLE_CHANGE]);
+
+      return;
+    } else {
+      changeRole({
+        organizationId: organizationId,
+        body: {
+          user_id: user_id,
+          role: selectedOption?.value,
+        },
       })
-      .catch((err) => {
-        toast.error(err?.data?.error || 'Failed to change role');
-      });
+        .unwrap()
+        .then(() => {
+          setSelectedRole(selectedOption);
+          setOpenChangeRoleDropdown(false);
+          setIsHoveredDropdown(false);
+          refetchAudiencesByOrganizationId();
+          toast.success(TOAST_MESSAGES.SUCCESS_AUDIENCE_ROLE_CHANGED);
+        })
+        .catch((err) => {
+          toast.error(err?.data?.error || TOAST_MESSAGES.FAILED_AUDIENCE_ROLE_CHANGED);
+        });
+    }
   };
 
   const handleOpenRemoveFromTeamPopup = () => {
@@ -68,20 +79,28 @@ const TeamMembersRole: FC<TeamMembersRolePropsType> = ({ value }) => {
     setIsOpenRemoveFromTeamPopup(false);
   };
 
-  const handleDeleteAudience = async () => {
-    try {
-      await deleteAudience({
+  const handleDeleteAudience = () => {
+    if (checkPermission) {
+      toast.error(PERMISSION_MESSAGES[PERMISSION_TYPES.DELETE]);
+
+      return;
+    } else {
+      deleteAudience({
         organizationId: organizationId,
         body: {
           user_id: user_id,
         },
-      }).unwrap();
-      handleCloseRemoveFromTeamPopup();
-      refetchAudiencesByOrganizationId();
-      toast.success('Audience deleted successfully');
-    } catch {
-      handleCloseRemoveFromTeamPopup();
-      toast.error('Failed to delete audience');
+      })
+        .unwrap()
+        .then(() => {
+          handleCloseRemoveFromTeamPopup();
+          refetchAudiencesByOrganizationId();
+          toast.success(TOAST_MESSAGES.SUCCESS_AUDIENCE_DELETED);
+        })
+        .catch((err) => {
+          handleCloseRemoveFromTeamPopup();
+          toast.error(err?.data?.error || TOAST_MESSAGES.FAILED_AUDIENCE_DELETED);
+        });
     }
   };
 
@@ -89,25 +108,29 @@ const TeamMembersRole: FC<TeamMembersRolePropsType> = ({ value }) => {
 
   return (
     <div className='w-full h-full text-left'>
-      <div className='relative w-fit'>
-        <AsyncDropdown
-          onOpen={handleOpenChangeRoleDropdown}
-          onClose={handleCloseChangeRoleDropdown}
-          isOpen={openChangeRoleDropdown}
-          onDelete={handleOpenRemoveFromTeamPopup}
-          onChange={(role) => handleRoleChange(role as TeamMemberAccessPrivilegesType)}
-          options={TEAM_MEMBERS_PRIVILEGES_LIST}
-          selectedValue={selectedRole}
-          defaultValue={role as TeamMemberAccessPrivilegesType}
-          showDelete
-          isHoveredDropdown={isHoveredDropdown}
-          setIsHoveredDropdown={setIsHoveredDropdown}
-          showSelectedIcon
-          parentWrapperClassName='pl-2'
-          wrapperClassName='w-[200px]'
-          selectedOptionClassName='bg-GRAY_100'
-        />
-      </div>
+      {checkPermission ? (
+        <div className='relative w-fit'>
+          <AsyncDropdown
+            onOpen={handleOpenChangeRoleDropdown}
+            onClose={handleCloseChangeRoleDropdown}
+            isOpen={openChangeRoleDropdown}
+            onDelete={handleOpenRemoveFromTeamPopup}
+            onChange={(role) => handleRoleChange(role as TeamMemberAccessPrivilegesType)}
+            options={TEAM_MEMBERS_PRIVILEGES_LIST}
+            selectedValue={selectedRole}
+            defaultValue={role as TeamMemberAccessPrivilegesType}
+            showDelete
+            showSelectedIcon
+            isHoveredDropdown={isHoveredDropdown}
+            setIsHoveredDropdown={setIsHoveredDropdown}
+            parentWrapperClassName='pl-2'
+            wrapperClassName='w-[200px]'
+            selectedOptionClassName='!bg-GRAY_100 !py-2.5'
+          />
+        </div>
+      ) : (
+        <span className='flex justify-between items-start f-12-400 text-GRAY_1000 pl-2 py-3 pr-2'>{role?.label}</span>
+      )}
       <RemoveFromTeamPopup
         isOpen={isOpenRemoveFromTeamPopup}
         onClose={handleCloseRemoveFromTeamPopup}
